Remove software floats

+++ backend/main/acv.h
@@ -1,12 +1,10 @@
 #ifndef ACV_H
 #define ACV_H
 
-#include "freertos/FreeRTOS.h"
-
 typedef struct {
     bool updated;
     int8_t coolant_temp;
-    float rpm;
+    uint16_t rpm;
     int8_t speed;
     int8_t intake_air_temp;
     float throttle_position;
@@ -15,7 +13,7 @@ typedef struct {
 extern acv_msg_t acv_msg;
 
 void acv_msg_coolant_temp(int8_t coolant_temp);
-void acv_msg_rpm(float rpm);
+void acv_msg_rpm(uint16_t rpm);
 void acv_msg_speed(int8_t speed);
 void acv_msg_intake_air_temp(int8_t intake_air_temp);
 void acv_msg_throttle_position(float throttle_position);
@@ -24,4 +22,4 @@ void acv_msg_string(char *buf, size_t size);
 
 void initialize_acv();
 
-#endif  // ACV_H
+#endif // ACV_H
